requirement.txt
json
pyperclip
pynput
pyscreenshot
pillow
requests
